fix(migrations): squash 24 migrations into single initial schema
Replaces 24 individual Alembic migrations with a single consolidated
0001_initial_schema.py. The old chain had a duplicate revision ID
(a1b2c3d4e5f6 used by both settings and impersonation) and a forked
head. Since the Postgres semantic layer is pre-production with no
deployed instances, squashing is the cleanest fix.

Single head: 0001. All 37 tables preserved with correct dependency
ordering, constraints, indexes, and foreign keys.
