--- src/dev_health_ops/alembic/versions/b2c3d4e5f6a7_add_user_org_membership_tables.py
@@ -1,118 +1,0 @@
-"""Add user, organization, and membership tables for Enterprise Edition.
-
-Revision ID: b2c3d4e5f6a7
-Revises: a1b2c3d4e5f6
-Create Date: 2026-01-30 09:00:00
-
-Creates tables for:
-- users: User accounts with authentication info
-- organizations: Multi-tenant organization containers
-- memberships: User-Organization relationships with roles
-"""
-
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects.postgresql import UUID
-
-
-revision = "b2c3d4e5f6a7"
-down_revision = "a1b2c3d4e5f6"
-branch_labels = None
-depends_on = None
-
-
-def upgrade() -> None:
-    op.create_table(
-        "users",
-        sa.Column("id", UUID(as_uuid=True), primary_key=True),
-        sa.Column("email", sa.Text(), nullable=False),
-        sa.Column("username", sa.Text(), nullable=True),
-        sa.Column("password_hash", sa.Text(), nullable=True),
-        sa.Column("full_name", sa.Text(), nullable=True),
-        sa.Column("avatar_url", sa.Text(), nullable=True),
-        sa.Column("auth_provider", sa.Text(), server_default="local"),
-        sa.Column("auth_provider_id", sa.Text(), nullable=True),
-        sa.Column("is_active", sa.Boolean(), nullable=False, server_default="true"),
-        sa.Column("is_verified", sa.Boolean(), nullable=False, server_default="false"),
-        sa.Column("is_superuser", sa.Boolean(), nullable=False, server_default="false"),
-        sa.Column("last_login_at", sa.DateTime(timezone=True), nullable=True),
-        sa.Column(
-            "created_at",
-            sa.DateTime(timezone=True),
-            nullable=False,
-            server_default=sa.text("CURRENT_TIMESTAMP"),
-        ),
-        sa.Column(
-            "updated_at",
-            sa.DateTime(timezone=True),
-            nullable=False,
-            server_default=sa.text("CURRENT_TIMESTAMP"),
-        ),
-    )
-    op.create_index("ix_users_email", "users", ["email"], unique=True)
-    op.create_index("ix_users_username", "users", ["username"], unique=True)
-    op.create_index(
-        "ix_users_auth_provider", "users", ["auth_provider", "auth_provider_id"]
-    )
-
-    op.create_table(
-        "organizations",
-        sa.Column("id", UUID(as_uuid=True), primary_key=True),
-        sa.Column("slug", sa.Text(), nullable=False),
-        sa.Column("name", sa.Text(), nullable=False),
-        sa.Column("description", sa.Text(), nullable=True),
-        sa.Column("settings", sa.JSON(), nullable=False, server_default="{}"),
-        sa.Column("tier", sa.Text(), nullable=False, server_default="'free'"),
-        sa.Column("stripe_customer_id", sa.Text(), nullable=True),
-        sa.Column("is_active", sa.Boolean(), nullable=False, server_default="true"),
-        sa.Column(
-            "created_at",
-            sa.DateTime(timezone=True),
-            nullable=False,
-            server_default=sa.text("CURRENT_TIMESTAMP"),
-        ),
-        sa.Column(
-            "updated_at",
-            sa.DateTime(timezone=True),
-            nullable=False,
-            server_default=sa.text("CURRENT_TIMESTAMP"),
-        ),
-    )
-    op.create_index("ix_organizations_slug", "organizations", ["slug"], unique=True)
-
-    op.create_table(
-        "memberships",
-        sa.Column("id", UUID(as_uuid=True), primary_key=True),
-        sa.Column("user_id", UUID(as_uuid=True), nullable=False),
-        sa.Column("org_id", UUID(as_uuid=True), nullable=False),
-        sa.Column("role", sa.Text(), nullable=False, server_default="'member'"),
-        sa.Column("invited_by_id", UUID(as_uuid=True), nullable=True),
-        sa.Column("joined_at", sa.DateTime(timezone=True), nullable=True),
-        sa.Column(
-            "created_at",
-            sa.DateTime(timezone=True),
-            nullable=False,
-            server_default=sa.text("CURRENT_TIMESTAMP"),
-        ),
-        sa.Column(
-            "updated_at",
-            sa.DateTime(timezone=True),
-            nullable=False,
-            server_default=sa.text("CURRENT_TIMESTAMP"),
-        ),
-        sa.ForeignKeyConstraint(["user_id"], ["users.id"], ondelete="CASCADE"),
-        sa.ForeignKeyConstraint(["org_id"], ["organizations.id"], ondelete="CASCADE"),
-        sa.ForeignKeyConstraint(["invited_by_id"], ["users.id"], ondelete="SET NULL"),
-    )
-    op.create_index("ix_memberships_user_id", "memberships", ["user_id"])
-    op.create_index("ix_memberships_org_id", "memberships", ["org_id"])
-    op.create_index("ix_memberships_org_role", "memberships", ["org_id", "role"])
-    op.create_unique_constraint(
-        "uq_membership_user_org", "memberships", ["user_id", "org_id"]
-    )
-
-
-def downgrade() -> None:
-    op.drop_table("memberships")
-    op.drop_table("organizations")
-    op.drop_table("users")
