--- src/dev_health_ops/alembic/versions/e5f6a7b8c9d0_add_audit_log_table.py
@@ -1,89 +1,0 @@
-"""Add audit_log table for Enterprise compliance.
-
-Revision ID: e5f6a7b8c9d0
-Revises: d4e5f6a7b8c9
-Create Date: 2026-01-30
-
-"""
-
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects.postgresql import UUID
-
-
-revision: str = "e5f6a7b8c9d0"
-down_revision: Union[str, None] = "d4e5f6a7b8c9"
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    op.create_table(
-        "audit_logs",
-        sa.Column("id", UUID(as_uuid=True), primary_key=True),
-        sa.Column(
-            "org_id",
-            UUID(as_uuid=True),
-            sa.ForeignKey("organizations.id", ondelete="CASCADE"),
-            nullable=False,
-        ),
-        sa.Column(
-            "user_id",
-            UUID(as_uuid=True),
-            sa.ForeignKey("users.id", ondelete="SET NULL"),
-            nullable=True,
-        ),
-        sa.Column("action", sa.Text(), nullable=False),
-        sa.Column("resource_type", sa.Text(), nullable=False),
-        sa.Column("resource_id", sa.Text(), nullable=False),
-        sa.Column("description", sa.Text(), nullable=True),
-        sa.Column("changes", sa.JSON(), nullable=True),
-        sa.Column("request_metadata", sa.JSON(), nullable=True),
-        sa.Column("status", sa.Text(), nullable=False, server_default="success"),
-        sa.Column("error_message", sa.Text(), nullable=True),
-        sa.Column(
-            "created_at",
-            sa.DateTime(timezone=True),
-            nullable=False,
-            server_default=sa.func.now(),
-        ),
-    )
-
-    op.create_index("ix_audit_logs_org_id", "audit_logs", ["org_id"])
-    op.create_index("ix_audit_logs_user_id", "audit_logs", ["user_id"])
-    op.create_index("ix_audit_logs_action", "audit_logs", ["action"])
-    op.create_index("ix_audit_logs_resource_type", "audit_logs", ["resource_type"])
-    op.create_index("ix_audit_logs_resource_id", "audit_logs", ["resource_id"])
-    op.create_index("ix_audit_logs_created_at", "audit_logs", ["created_at"])
-    op.create_index("ix_audit_logs_org_created", "audit_logs", ["org_id", "created_at"])
-    op.create_index(
-        "ix_audit_logs_user_created", "audit_logs", ["user_id", "created_at"]
-    )
-    op.create_index(
-        "ix_audit_logs_resource", "audit_logs", ["resource_type", "resource_id"]
-    )
-    op.create_index(
-        "ix_audit_logs_action_created", "audit_logs", ["action", "created_at"]
-    )
-    op.create_index(
-        "ix_audit_logs_org_action_created",
-        "audit_logs",
-        ["org_id", "action", "created_at"],
-    )
-
-
-def downgrade() -> None:
-    op.drop_index("ix_audit_logs_org_action_created", table_name="audit_logs")
-    op.drop_index("ix_audit_logs_action_created", table_name="audit_logs")
-    op.drop_index("ix_audit_logs_resource", table_name="audit_logs")
-    op.drop_index("ix_audit_logs_user_created", table_name="audit_logs")
-    op.drop_index("ix_audit_logs_org_created", table_name="audit_logs")
-    op.drop_index("ix_audit_logs_created_at", table_name="audit_logs")
-    op.drop_index("ix_audit_logs_resource_id", table_name="audit_logs")
-    op.drop_index("ix_audit_logs_resource_type", table_name="audit_logs")
-    op.drop_index("ix_audit_logs_action", table_name="audit_logs")
-    op.drop_index("ix_audit_logs_user_id", table_name="audit_logs")
-    op.drop_index("ix_audit_logs_org_id", table_name="audit_logs")
-    op.drop_table("audit_logs")
