--- src/dev_health_ops/alembic/versions/o5j6k7l8m9n0_add_subscriptions.py
@@ -1,116 +1,0 @@
-"""Add subscriptions and subscription_events tables.
-
-Revision ID: o5j6k7l8m9n0
-Revises: n4i5j6k7l8m9
-Create Date: 2026-02-24 20:00:00
-"""
-
-from typing import Union
-
-import sqlalchemy as sa
-from alembic import op
-from sqlalchemy.dialects import postgresql
-
-
-revision: str = "o5j6k7l8m9n0"
-down_revision: Union[str, None] = "n4i5j6k7l8m9"
-branch_labels: Union[str, None] = None
-depends_on: Union[str, None] = None
-
-
-def upgrade() -> None:
-    op.create_table(
-        "subscriptions",
-        sa.Column("id", postgresql.UUID(), primary_key=True),
-        sa.Column("org_id", postgresql.UUID(), nullable=False),
-        sa.Column("billing_plan_id", postgresql.UUID(), nullable=False),
-        sa.Column("billing_price_id", postgresql.UUID(), nullable=False),
-        sa.Column("stripe_subscription_id", sa.Text(), nullable=False),
-        sa.Column("stripe_customer_id", sa.Text(), nullable=False),
-        sa.Column("status", sa.Text(), nullable=False),
-        sa.Column("current_period_start", sa.DateTime(timezone=True), nullable=False),
-        sa.Column("current_period_end", sa.DateTime(timezone=True), nullable=False),
-        sa.Column("cancel_at_period_end", sa.Boolean(), server_default="false"),
-        sa.Column("canceled_at", sa.DateTime(timezone=True), nullable=True),
-        sa.Column("trial_start", sa.DateTime(timezone=True), nullable=True),
-        sa.Column("trial_end", sa.DateTime(timezone=True), nullable=True),
-        sa.Column("metadata", sa.JSON(), server_default="{}", nullable=False),
-        sa.Column(
-            "created_at",
-            sa.DateTime(timezone=True),
-            nullable=False,
-            server_default=sa.text("now()"),
-        ),
-        sa.Column(
-            "updated_at",
-            sa.DateTime(timezone=True),
-            nullable=False,
-            server_default=sa.text("now()"),
-        ),
-        sa.ForeignKeyConstraint(["org_id"], ["organizations.id"], ondelete="CASCADE"),
-        sa.ForeignKeyConstraint(
-            ["billing_plan_id"],
-            ["billing_plans.id"],
-            ondelete="RESTRICT",
-        ),
-        sa.ForeignKeyConstraint(
-            ["billing_price_id"],
-            ["billing_prices.id"],
-            ondelete="RESTRICT",
-        ),
-        sa.UniqueConstraint(
-            "stripe_subscription_id", name="uq_subscriptions_stripe_id"
-        ),
-    )
-    op.create_index("ix_subscriptions_org_id", "subscriptions", ["org_id"])
-    op.create_index(
-        "ix_subscriptions_customer_id",
-        "subscriptions",
-        ["stripe_customer_id"],
-    )
-    op.create_index(
-        "ix_subscriptions_stripe_subscription_id",
-        "subscriptions",
-        ["stripe_subscription_id"],
-    )
-
-    op.create_table(
-        "subscription_events",
-        sa.Column("id", postgresql.UUID(), primary_key=True),
-        sa.Column("subscription_id", postgresql.UUID(), nullable=False),
-        sa.Column("stripe_event_id", sa.Text(), nullable=False),
-        sa.Column("event_type", sa.Text(), nullable=False),
-        sa.Column("previous_status", sa.Text(), nullable=True),
-        sa.Column("new_status", sa.Text(), nullable=False),
-        sa.Column("payload", sa.JSON(), server_default="{}", nullable=False),
-        sa.Column(
-            "processed_at",
-            sa.DateTime(timezone=True),
-            nullable=False,
-            server_default=sa.text("now()"),
-        ),
-        sa.ForeignKeyConstraint(
-            ["subscription_id"], ["subscriptions.id"], ondelete="CASCADE"
-        ),
-        sa.UniqueConstraint(
-            "stripe_event_id", name="uq_subscription_events_stripe_event"
-        ),
-    )
-    op.create_index(
-        "ix_subscription_events_subscription_id",
-        "subscription_events",
-        ["subscription_id"],
-    )
-
-
-def downgrade() -> None:
-    op.drop_index(
-        "ix_subscription_events_subscription_id",
-        table_name="subscription_events",
-    )
-    op.drop_table("subscription_events")
-
-    op.drop_index("ix_subscriptions_stripe_subscription_id", table_name="subscriptions")
-    op.drop_index("ix_subscriptions_customer_id", table_name="subscriptions")
-    op.drop_index("ix_subscriptions_org_id", table_name="subscriptions")
-    op.drop_table("subscriptions")
