--- src/dev_health_ops/alembic/versions/p6k7l8m9n0o1_add_invoices.py
@@ -1,87 +1,0 @@
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects.postgresql import UUID
-
-
-revision = "p6k7l8m9n0o1"
-down_revision = "o5j6k7l8m9n0"
-branch_labels = None
-depends_on = None
-
-
-def upgrade() -> None:
-    op.create_table(
-        "invoices",
-        sa.Column("id", UUID(as_uuid=True), nullable=False),
-        sa.Column("org_id", UUID(as_uuid=True), nullable=False),
-        sa.Column("subscription_id", UUID(as_uuid=True), nullable=True),
-        sa.Column("stripe_invoice_id", sa.Text(), nullable=False),
-        sa.Column("stripe_customer_id", sa.Text(), nullable=False),
-        sa.Column("status", sa.Text(), nullable=False),
-        sa.Column("amount_due", sa.Integer(), nullable=False),
-        sa.Column("amount_paid", sa.Integer(), nullable=False, server_default="0"),
-        sa.Column(
-            "amount_remaining",
-            sa.Integer(),
-            nullable=False,
-            server_default="0",
-        ),
-        sa.Column("currency", sa.Text(), nullable=False, server_default="usd"),
-        sa.Column("period_start", sa.DateTime(timezone=True), nullable=True),
-        sa.Column("period_end", sa.DateTime(timezone=True), nullable=True),
-        sa.Column("hosted_invoice_url", sa.Text(), nullable=True),
-        sa.Column("pdf_url", sa.Text(), nullable=True),
-        sa.Column("payment_intent_id", sa.Text(), nullable=True),
-        sa.Column("finalized_at", sa.DateTime(timezone=True), nullable=True),
-        sa.Column("paid_at", sa.DateTime(timezone=True), nullable=True),
-        sa.Column("voided_at", sa.DateTime(timezone=True), nullable=True),
-        sa.Column("attempt_count", sa.Integer(), nullable=False, server_default="0"),
-        sa.Column("metadata", sa.JSON(), nullable=False, server_default="{}"),
-        sa.Column(
-            "created_at",
-            sa.DateTime(timezone=True),
-            nullable=False,
-            server_default=sa.text("now()"),
-        ),
-        sa.Column(
-            "updated_at",
-            sa.DateTime(timezone=True),
-            nullable=False,
-            server_default=sa.text("now()"),
-        ),
-        sa.ForeignKeyConstraint(["org_id"], ["organizations.id"]),
-        sa.ForeignKeyConstraint(["subscription_id"], ["subscriptions.id"]),
-        sa.PrimaryKeyConstraint("id"),
-        sa.UniqueConstraint("stripe_invoice_id"),
-    )
-    op.create_index("ix_invoices_org_id", "invoices", ["org_id"])
-    op.create_index("ix_invoices_subscription_id", "invoices", ["subscription_id"])
-    op.create_index("ix_invoices_stripe_invoice_id", "invoices", ["stripe_invoice_id"])
-    op.create_index(
-        "ix_invoices_stripe_customer_id", "invoices", ["stripe_customer_id"]
-    )
-
-    op.create_table(
-        "invoice_line_items",
-        sa.Column("id", UUID(as_uuid=True), nullable=False),
-        sa.Column("invoice_id", UUID(as_uuid=True), nullable=False),
-        sa.Column("stripe_line_item_id", sa.Text(), nullable=True),
-        sa.Column("description", sa.Text(), nullable=True),
-        sa.Column("amount", sa.Integer(), nullable=False),
-        sa.Column("quantity", sa.Integer(), nullable=False, server_default="1"),
-        sa.Column("period_start", sa.DateTime(timezone=True), nullable=True),
-        sa.Column("period_end", sa.DateTime(timezone=True), nullable=True),
-        sa.Column("stripe_price_id", sa.Text(), nullable=True),
-        sa.ForeignKeyConstraint(["invoice_id"], ["invoices.id"], ondelete="CASCADE"),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_index(
-        "ix_invoice_line_items_invoice_id",
-        "invoice_line_items",
-        ["invoice_id"],
-    )
-
-
-def downgrade() -> None:
-    op.drop_table("invoice_line_items")
-    op.drop_table("invoices")
