--- src/dev_health_ops/alembic/versions/s9n0o1p2q3r4_add_refresh_tokens_table.py
@@ -1,70 +1,0 @@
-"""Add refresh_tokens table for server-side refresh token management.
-
-Revision ID: s9n0o1p2q3r4
-Revises: r8m9n0o1p2q3
-Create Date: 2026-02-26
-"""
-
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects.postgresql import UUID
-
-
-revision: str = "s9n0o1p2q3r4"
-down_revision: Union[str, None] = "r8m9n0o1p2q3"
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    op.create_table(
-        "refresh_tokens",
-        sa.Column("id", UUID(as_uuid=True), primary_key=True),
-        sa.Column(
-            "user_id",
-            UUID(as_uuid=True),
-            sa.ForeignKey("users.id", ondelete="CASCADE"),
-            nullable=False,
-        ),
-        sa.Column(
-            "org_id",
-            UUID(as_uuid=True),
-            sa.ForeignKey("organizations.id", ondelete="CASCADE"),
-            nullable=False,
-        ),
-        sa.Column("token_hash", sa.Text(), nullable=False),
-        sa.Column("family_id", UUID(as_uuid=True), nullable=False),
-        sa.Column("expires_at", sa.DateTime(timezone=True), nullable=False),
-        sa.Column("revoked_at", sa.DateTime(timezone=True), nullable=True),
-        sa.Column("replaced_by_hash", sa.Text(), nullable=True),
-        sa.Column("ip_address", sa.Text(), nullable=True),
-        sa.Column("user_agent", sa.Text(), nullable=True),
-        sa.Column(
-            "created_at",
-            sa.DateTime(timezone=True),
-            nullable=False,
-            server_default=sa.func.now(),
-        ),
-        sa.UniqueConstraint("token_hash", name="uq_refresh_tokens_token_hash"),
-    )
-
-    op.create_index("ix_refresh_tokens_user_id", "refresh_tokens", ["user_id"])
-    op.create_index("ix_refresh_tokens_token_hash", "refresh_tokens", ["token_hash"])
-    op.create_index("ix_refresh_tokens_family_id", "refresh_tokens", ["family_id"])
-    op.create_index(
-        "ix_refresh_tokens_user_family",
-        "refresh_tokens",
-        ["user_id", "family_id"],
-    )
-    op.create_index("ix_refresh_tokens_expires", "refresh_tokens", ["expires_at"])
-
-
-def downgrade() -> None:
-    op.drop_index("ix_refresh_tokens_expires", table_name="refresh_tokens")
-    op.drop_index("ix_refresh_tokens_user_family", table_name="refresh_tokens")
-    op.drop_index("ix_refresh_tokens_family_id", table_name="refresh_tokens")
-    op.drop_index("ix_refresh_tokens_token_hash", table_name="refresh_tokens")
-    op.drop_index("ix_refresh_tokens_user_id", table_name="refresh_tokens")
-    op.drop_table("refresh_tokens")
