--- src/dev_health_ops/alembic/versions/u1b2c3d4e5f6_add_password_reset_tokens_table.py
@@ -1,75 +1,0 @@
-"""Add password_reset_tokens table.
-
-Revision ID: u1b2c3d4e5f6
-Revises: t0a1b2c3d4e6
-Create Date: 2026-02-26
-"""
-
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects.postgresql import UUID
-
-
-revision: str = "u1b2c3d4e5f6"
-down_revision: Union[str, None] = "t0a1b2c3d4e6"
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    op.create_table(
-        "password_reset_tokens",
-        sa.Column("id", UUID(as_uuid=True), primary_key=True),
-        sa.Column(
-            "user_id",
-            UUID(as_uuid=True),
-            sa.ForeignKey("users.id", ondelete="CASCADE"),
-            nullable=False,
-        ),
-        sa.Column("token_hash", sa.Text(), nullable=False),
-        sa.Column("expires_at", sa.DateTime(timezone=True), nullable=False),
-        sa.Column(
-            "created_at",
-            sa.DateTime(timezone=True),
-            nullable=False,
-            server_default=sa.func.now(),
-        ),
-        sa.UniqueConstraint(
-            "token_hash",
-            name="uq_password_reset_tokens_token_hash",
-        ),
-    )
-
-    op.create_index(
-        "ix_password_reset_tokens_user_id",
-        "password_reset_tokens",
-        ["user_id"],
-    )
-    op.create_index(
-        "ix_password_reset_tokens_token_hash",
-        "password_reset_tokens",
-        ["token_hash"],
-    )
-    op.create_index(
-        "ix_password_reset_tokens_user_expires",
-        "password_reset_tokens",
-        ["user_id", "expires_at"],
-    )
-
-
-def downgrade() -> None:
-    op.drop_index(
-        "ix_password_reset_tokens_user_expires",
-        table_name="password_reset_tokens",
-    )
-    op.drop_index(
-        "ix_password_reset_tokens_token_hash",
-        table_name="password_reset_tokens",
-    )
-    op.drop_index(
-        "ix_password_reset_tokens_user_id",
-        table_name="password_reset_tokens",
-    )
-    op.drop_table("password_reset_tokens")
